Added a 4th cone to the Red Left
Added the 4th cone to the trajectories
Adjusted the cone height delta as the hand is not coming down quite far enough for 4th cone
Parking issues with reading the sleeve, but Nate's code was able to read.  Need to check to see if we can replicate how his is reading. Still working on heading issues with parking so it doesn't turn more than it needs to