Update AutonomousRedLeftTestForLow.java
Issues with reading signal sleeve:  Changed tensor to tfod to match Nate's code.  Ending up commenting out the while statement for tensor flow which now gives us more consistent readings.  We can get 4 cones and park in position 1, but parking in 2 or 3 is very inconsistent due to running out of time.  Plan is to get 3 cones and park for this weekend's qualifier - 30 pts.